Digital-Krishi-Officer-main/AGRIVERSE-24-7/backend/requirements.txt
fastapi==0.101.1
uvicorn==0.23.2
python-multipart==0.0.6

langchain==0.1.16
langchain-community==0.0.38

faiss-cpu
sentence-transformers==2.2.2

requests==2.31.0
pillow==10.4.0
sqlalchemy==1.4.49
googletrans
python-dotenv==1.0.0
asyncpg==0.27.0
greenlet==2.0.2
langdetect==1.0.9
mtranslate==1.8
groq==0.4.2
deep-translator==1.11.4
